Add and edit index.html and style.css
The index.html and style.css files of the landing page has been added
to the repository. Progress of the landing page design is at around
40%, with the top half of the website completed.

As per the recommendation on The Odin Project, the approach to
designing the landing page is of a modular style from top to bottom
of the landing page, with the html code for each portion completed
before css styling begins. This makes the coding process easier to
tackle.